requirements.txt
requests
easygui
discord.py
colorama
